backend/requirements.txt
Flask
python-dotenv
deepgram-sdk
google-generativeai
flask-cors
elevenlabs 